Move survey files to folder
