+++ src/js/project/ReviewForm.js
@@ -5,8 +5,8 @@ import {ImageryReview} from "./ImagerySelection";
 import {OverviewReview} from "./Overview";
 import {AOIReview, PlotReview} from "./PlotDesign";
 import {SampleReview} from "./SampleDesign";
-import SurveyCardList from "./SurveyCardList";
-import {SurveyRulesList} from "./SurveyRules";
+import SurveyCardList from "../survey/SurveyCardList";
+import {SurveyRulesList} from "../survey/SurveyRules";
 import SvgIcon from "../components/svg/SvgIcon";
 import UserAssignmentReview from "./UserAssignmentReview";
 import QAQCReview from "./QAQCReview";
